EES-4775 Move redirectPages code into middleware.ts
